backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
mss>=9.0.0
paddleocr>=2.7.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pillow>=10.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
